Refactor file structure to share common golang code within module
