+++ lambda_golang/cmd/stories/main.go
