+++ lambda_golang/go.mod
@@ -1,0 +1,15 @@
+module github.com/rivernews/media-literacy
+
+go 1.16
+
+require (
+	github.com/PuerkitoBio/goquery v1.7.1
+	github.com/aws/aws-lambda-go v1.26.0
+	github.com/aws/aws-sdk-go-v2 v1.9.0
+	github.com/aws/aws-sdk-go-v2/config v1.7.0
+	github.com/aws/aws-sdk-go-v2/feature/s3/manager v1.5.0
+	github.com/aws/aws-sdk-go-v2/service/s3 v1.14.0
+	github.com/aws/aws-sdk-go-v2/service/sqs v1.8.0
+	github.com/rivernews/GoTools v0.1.6
+	golang.org/x/net v0.0.0-20210825183410-e898025ed96a
+)
