+++ lambda_golang/pkg/newssite/utilities.go
@@ -1,4 +1,4 @@
-package main
+package newssite
 
 import (
 	"strings"
