--- scraper_lambda/landing/go.mod
@@ -1,14 +1,0 @@
-module github.com/rivernews/media-literacy
-
-go 1.16
-
-require (
-	github.com/PuerkitoBio/goquery v1.7.1
-	github.com/aws/aws-lambda-go v1.25.0
-	github.com/aws/aws-sdk-go-v2 v1.8.0
-	github.com/aws/aws-sdk-go-v2/config v1.6.0
-	github.com/aws/aws-sdk-go-v2/feature/s3/manager v1.4.0
-	github.com/aws/aws-sdk-go-v2/service/s3 v1.12.0
-	github.com/rivernews/GoTools v0.1.6
-	golang.org/x/net v0.0.0-20210726213435-c6fcb2dbf985
-)
