--- scraper_lambda/stories/go.mod
@@ -1,11 +1,0 @@
-module github.com/rivernews/media-literacy/stories
-
-go 1.16
-
-require (
-	github.com/aws/aws-lambda-go v1.26.0
-	github.com/aws/aws-sdk-go-v2 v1.9.0
-	github.com/aws/aws-sdk-go-v2/config v1.7.0
-	github.com/aws/aws-sdk-go-v2/service/sqs v1.8.0
-	github.com/rivernews/GoTools v0.1.6
-)
