--- scraper_lambda/story/go.mod
@@ -1,8 +1,0 @@
-module github.com/rivernews/media-literacy/story
-
-go 1.16
-
-require (
-	github.com/aws/aws-lambda-go v1.26.0
-	github.com/rivernews/GoTools v0.1.6
-)
